% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_assignments}
\alias{compute_assignments}
\title{Derive crisp assignments for given set of prototypes}
\usage{
compute_assignments(X, W, dist_fun)
}
\arguments{
\item{X}{The N x n design matrix, N observations and n (preferably) scaled features}

\item{W}{The K x n prototype matrix returned by the neural gas algorithm}

\item{dist_fun}{The distance or dissimilarity function used.}
}
\value{
A vector of length N, containing the prototype index for each of the N observations.
}
\description{
Derive crisp assignments for given set of prototypes
}
