% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_euclidean.R
\name{dists_xi_euclidean}
\alias{dists_xi_euclidean}
\title{Compute euclidean distances between some data vector x_k and prototypes}
\usage{
dists_xi_euclidean(W, x_i)
}
\arguments{
\item{W}{The K x n prototype matrix}

\item{x_k}{The (scaled) k-th data vector}
}
\value{
K-dimensional vector containing euclidean distances.
}
\description{
Compute euclidean distances between some data vector x_k and prototypes
}
