% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralgas.R
\name{neuralgas}
\alias{neuralgas}
\title{Run Neural Gas algorithm}
\usage{
neuralgas(
  X,
  K = NULL,
  W = NULL,
  lambda_init = NULL,
  lambda_stop = 0.001,
  learn_rate = 0.1,
  max_iter = nrow(X) * 100,
  dists_fun = dists_xi_euclidean,
  update_fun = update_euclidean,
  verbose = T,
  refresh = 250
)
}
\arguments{
\item{X}{An N x n data matrix of N observations with n (preferably scaled) features.}

\item{K}{Number of prototypes to initialize}

\item{W}{Optional K x n matrix of prototypes. When provided, the K argument is ignored.}

\item{lambda_init}{Initial adaption strength in the first iteration.}

\item{lambda_stop}{Target adaption strength, effective in the final iteration}

\item{learn_rate}{Constant learn rate. Defaults to 0.1.}

\item{max_iter}{Number of iterations to run. Defaults to 100 x N.}

\item{dists_fun}{The funtion to compute K distances / dissimilarities given a data vector and the prototype matrix.}

\item{update_fun}{The funtion to compute the update for each of the K prototypes.}

\item{verbose}{Print life signs.}

\item{refresh}{Refresh life signs printout after certain number of iterations. Defaults to 250 iterations.}
}
\value{
A list, similar to returned result of the base R kmeans implementation: A matrix of K prototypes, crisp assignment per data point, and number of assignments per prototype.
}
\description{
Run Neural Gas algorithm
}
