% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_expose.R
\name{NGConvergence}
\alias{NGConvergence}
\title{Convergence for NeuralGas Objects}
\description{
\strong{Note:} This is not a function, merely a description of the convergence criteria used in the \code{NGBatch} and \code{NGOnline} functions.
}
\details{
NeuralGas learning is terminated by the first occurrence of the following convergence criteria reaching their user-supplied tolerances: 

\itemize{
\item The training age (number of learning epochs performed) exceeding the parameter \code{max_epochs} 

\item The \code{tol_delBMU} AND \code{tol_delQE} tolerances being met for three consecutive epochs. 
These criteria, described below, measure the stability of the vector quantization over training time.  
}

\code{delBMU} reports the percentage of training vectors whose BMU has changed from one epoch to the next. 
For example, \code{delBMU < tol_delBMU = 1} means that 
less than 1\% of the data have changed their BMU from epoch \code{t-1} to \code{t}. 

\code{delQE} reports the average (absolute) percentage change in each datum's quantization error (QE) from one epoch to the next. 
For example, \code{delQE < tol_delQE = 1} means that, on average, each training vector's QE has changed by less than 1\% from epoch \code{t-1} to \code{t}.  

The default value \code{max_epochs = -1} indicates convergence (training cessation) is measured ONLY by \code{delBMU} and \code{delQE}.  
Thus, to train for a fixed number of epochs \code{E}, set \code{max_epochs = E} and \code{tol_delBMU = tol_delQE = 0}. 

For online learning an epoch is equivalent to \code{N} learning iterations, where \code{N} = number of training vectors. 

During training early termination can be achieved by user interruption (typing CTRL-C in the R console window).  
Upon detecting early termination, both \code{NGBatch} and \code{NGOnline} will return the current state of the network 
(existing values of the prototypes and any training history logged).
}
