% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_expose.R
\name{NGWInitialization}
\alias{NGWInitialization}
\title{Prototype Initialization for NeuralGas Objects}
\description{
\strong{Note:} This is not a function, merely a description of the types of prototype initialization understood by the \code{NGBatch} and \code{NGOnline} functions.
}
\details{
Both \code{NGBatch} and \code{NGOnline} require the input parameter \code{W} which describes both 
the number of prototypes in the network, and how they are initialized.  As such, the value supplied for \code{W} can take 
one of the three following forms: 

\describe{
\item{a single number}{giving the number of prototypes in the network. 
In this case, prototypes are initialized randomly and uniformly in the range of training data \code{X}.}

\item{a length=2 vector}{giving the number of prototypes (first element) and the random seed used to initialize them (second element). 
Use this option for random prototype initialization with a fixed seed for reproducibility.}

\item{a matrix}{giving the initial prototypes in its rows. ncols(W) should = ncol(X), and the number of prototypes in the network is set equal to \code{nrow(W)}
If initial prototypes are given they should occupy the same range as \code{X}, 
as they will be linearly scaled internally from \[max(X),min(X)\] to \[0,1\] prior to learning.}
}
}
