% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_euclidean.R
\name{update_euclidean}
\alias{update_euclidean}
\title{The Neural Gas update rule for euclidean vector space-embedded data and prototypes}
\usage{
update_euclidean(x_i, W, learn_rate, h)
}
\arguments{
\item{x_i}{A (prefereably scaled) n-dimensional data vector}

\item{W}{The K x n prototype matrix}

\item{learn_rate}{Learn rate of the algorithm.}

\item{h}{The neighborhood-based adaption strength, given the dissimilarity ranks of the prototypes.}
}
\value{
The K x n  matrix of the K prototypes.
}
\description{
The Neural Gas update rule for euclidean vector space-embedded data and prototypes
}
